+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v0.6.0 (in development)
 -----------------------
 - Add URL of GitHub repository to `--help` output
 - Increased MSRV to 1.85
+- Always show magnet display name next to info hash in log messages
 
 v0.5.0 (2025-05-20)
 -------------------